refactor(graph): extract the source indexer + shared link/tag helpers (#1624)
Second per-format split of indexers.ts (arch D1). Source is more coupled than
excerpt — indexSourceBody reuses the note wiki-link/tag pipeline — so this also
relocates those shared helpers to the leaf so the split stays acyclic.

- index-helpers.ts (leaf) gains the shared body-indexing helpers used by both
  note and source indexing: `LinkResolveCtx` + `buildLinkResolveCtx` +
  `resolveLinkTarget` (wiki-link resolution), `ensureTag`, and
  `flattenFrontmatterStrings`. Pure relocation — the note-indexing hot path now
  imports them from the leaf.
- New `graph/indexers/source.ts`: indexSource, indexSourceBody, removeSource,
  parseSourceIdFromPath + the `walkAndIndexSources` walker (moved with it so
  indexAllNotes imports it, not the reverse — no cycle).
- indexers.ts imports the shared helpers + the walker; the facade re-exports the
  source functions from the new module. External callers unchanged.
  indexers.ts: 1559 → 1358 lines.

Behaviour-preserving — the full graph suite (440) and the sources suite (434)
are green; lint clean.

Follow-ups: tables and the note-file-type (.ttl/.csv/.py) indexers.
